- Exported RenderMode_t as RenderMode (forwarded to entities.constants).
- Fixed a no to-python converter error.

- Fixed a TypeError when a wrapped constant is not defined for a specific engine/game.

+++ src/core/modules/entities/csgo/entities_constants_wrap_python.h
@@ -61,9 +61,9 @@ void export_engine_specific_solid_flags(T SolidFlags)
 template<class T>
 void export_engine_specific_entity_effects(T _constants)
 {
-	_constants.attr("EF_MARKED_FOR_FAST_REFLECTION") = EF_MARKED_FOR_FAST_REFLECTION;
-	_constants.attr("EF_NOSHADOWDEPTH") = EF_NOSHADOWDEPTH;
-	_constants.attr("EF_NOFLASHLIGHT") = EF_NOFLASHLIGHT;
+	_constants.attr("EF_MARKED_FOR_FAST_REFLECTION") = (long)EF_MARKED_FOR_FAST_REFLECTION;
+	_constants.attr("EF_NOSHADOWDEPTH") = (long)EF_NOSHADOWDEPTH;
+	_constants.attr("EF_NOFLASHLIGHT") = (long)EF_NOFLASHLIGHT;
 }
